+++ host-agent/src/handlers/mod.rs
@@ -1,3 +1,3 @@
 pub mod files;
-//pub mod system;
+pub mod network;
 //pub mod process;